requirements.txt
streamlit
fastapi
uvicorn
pydantic
python-multipart
requests
numpy
pandas
PyMuPDF
sentence-transformers
faiss-cpu
transformers
torch
scikit-learn
langchain
python-docx